+++ src/main/java/gaji/service/domain/roomBoard/converter/RoomPostConverter.java
@@ -1,8 +1,8 @@
-package gaji.service.domain.roomPost.converter;
+package gaji.service.domain.roomBoard.converter;
 
-import gaji.service.domain.roomPost.entity.RoomBoard;
-import gaji.service.domain.roomPost.entity.RoomPost;
-import gaji.service.domain.roomPost.web.dto.RoomPostRequestDto;
+import gaji.service.domain.roomBoard.entity.RoomBoard;
+import gaji.service.domain.roomBoard.entity.RoomPost;
+import gaji.service.domain.roomBoard.web.dto.RoomPostRequestDto;
 import gaji.service.domain.user.entity.User;
 
 import java.time.LocalDateTime;
