+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPostBookmark.java
@@ -1,7 +1,6 @@
-package gaji.service.domain.roomPost.entity;
+package gaji.service.domain.roomBoard.entity;
 
 import gaji.service.domain.user.entity.User;
-import gaji.service.domain.enums.ReportPostTypeEnum;
 import jakarta.persistence.*;
 import lombok.AccessLevel;
 import lombok.Getter;
@@ -10,7 +9,8 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class RoomPostReport {
+public class RoomPostBookmark {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
@@ -20,11 +20,6 @@ public class RoomPostReport {
     private User user;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
+    @JoinColumn(name = "room_id")
     private RoomPost roomPost;
-
-    @Enumerated(EnumType.STRING)
-    private ReportPostTypeEnum reportPostTypeEnum;
-
-    private String description;
 }
