+++ src/main/java/gaji/service/domain/roomBoard/entity/RoomPostLikes.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.roomPost.entity;
+package gaji.service.domain.roomBoard.entity;
 
 import gaji.service.domain.user.entity.User;
 import jakarta.persistence.*;
