+++ src/main/java/gaji/service/domain/roomBoard/service/RoomPostCommandService.java
@@ -1,0 +1,9 @@
+package gaji.service.domain.roomBoard.service;
+
+import gaji.service.domain.roomBoard.entity.RoomPost;
+import gaji.service.domain.roomBoard.web.dto.RoomPostRequestDto;
+
+public interface RoomPostCommandService {
+    RoomPost createRoomPost(Long roomId, Long userId, RoomPostRequestDto.RoomPostDto requestDto);
+
+}
