+++ src/main/java/gaji/service/domain/roomBoard/web/dto/RoomPostRequestDto.java
@@ -1,4 +1,4 @@
-package gaji.service.domain.roomPost.web.dto;
+package gaji.service.domain.roomBoard.web.dto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
