--- src/main/java/gaji/service/domain/roomPost/service/RoomPostCommandService.java
@@ -1,10 +1,0 @@
-package gaji.service.domain.roomPost.service;
-
-import gaji.service.domain.roomPost.entity.RoomPost;
-import gaji.service.domain.roomPost.web.dto.RoomPostRequestDto;
-import jakarta.transaction.Transactional;
-
-public interface RoomPostCommandService {
-    RoomPost createRoomPost(Long roomId, Long userId, RoomPostRequestDto.RoomPostDto requestDto);
-
-}
