+++ src/main/java/gaji/service/domain/user/entity/User.java
@@ -13,7 +13,7 @@ import gaji.service.domain.recruit.entity.SearchKeyword;
 import gaji.service.domain.room.entity.Room;
 import gaji.service.domain.room.entity.RoomEvent;
 import gaji.service.domain.room.entity.VoiceChatUser;
-import gaji.service.domain.roomPost.entity.*;
+import gaji.service.domain.roomBoard.entity.*;
 import gaji.service.domain.studyMate.entity.*;
 import gaji.service.oauth2.dto.TransferUserDTO;
 import jakarta.persistence.*;
